Unique serial numbers
Firefox raises an error for certificates with the same serial numbers,
issued by the same CA.  This patch uses a unique serial number for each
certificate.

(Certificate generation is no longer deterministic.  A better patch
might make use of something like the mtime on the Cakefile?  It would
probably have to be something like that as it has to be monotonic.)